Rename `#program` to `#programs`
The other source directories have names in plural.
