+++ src/main-cli.ts
@@ -1,7 +1,7 @@
 #!/usr/bin/env node
 
 import process, { argv } from "node:process"
-import { cliProgram } from "#program/cli/CliProgram.ts"
+import { cliProgram } from "#programs/cli/CliProgram.ts"
 import type { ExitCode } from "#types/ExitCode.ts"
 
 const exitCode: ExitCode = await cliProgram(argv.slice(2))
