+++ src/programs/cli/CliProgram.tests.ts
@@ -1,7 +1,7 @@
 import { mockUpdraftVersion } from "#utilities/version/UpdraftVersion.mocks.ts"
 import { beforeEach, describe, expect, it } from "vitest"
-import { cliProgram } from "#program/cli/CliProgram.ts"
-import { usageInstructions } from "#program/cli/UsageInstructions.ts"
+import { cliProgram } from "#programs/cli/CliProgram.ts"
+import { usageInstructions } from "#programs/cli/UsageInstructions.ts"
 import { EXIT_CODE_SUCCESS, type ExitCode } from "#types/ExitCode.ts"
 import type { SemanticVersionString } from "#types/SemanticVersionString.ts"
 import { readMatchingFiles, writeFiles } from "#utilities/files/FileSystem.ts"
